# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NexoShop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get cart items
      responses:
        "200":
          description: Cart items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartSummary"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartSummary"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear cart
      responses:
        "200":
          description: Cart cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartSummary"
  /cart/{productId}:
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartSummary"
  /cart/validate-coupon:
    post:
      operationId: validateCoupon
      tags: [cart]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateCouponBody"
      responses:
        "200":
          description: Coupon validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidation"
  /cart/checkout:
    post:
      operationId: checkout
      tags: [cart]
      summary: Checkout (purchase cart)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutBody"
      responses:
        "200":
          description: Checkout result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResult"
        "400":
          description: Insufficient balance
...
  /orders/buy:
    post:
      operationId: buyProduct
      tags: [orders]
      summary: Buy a single product directly
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyProductBody"
      responses:
        "200":
          description: Purchase result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Insufficient balance
...
  /loyalty/convert:
    post:
      operationId: convertPoints
      tags: [loyalty]
      summary: Convert loyalty points to balance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConvertPointsBody"
      responses:
        "200":
          description: Conversion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConvertPointsResult"
        "400":
          description: Insufficient points
...
components:
  schemas:
...
    CartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productEmoji:
          type: string
        price:
          type: number
        quantity:
          type: integer
        deliveryType:
          type: string
      required: [id, productId, productName, productEmoji, price, quantity, deliveryType]
    CartSummary:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        couponCode:
          type: ["string", "null"]
        itemCount:
          type: integer
      required: [items, subtotal, discount, total, itemCount]
    AddToCartBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required: [productId]
    ValidateCouponBody:
      type: object
      properties:
        code:
          type: string
        cartTotal:
          type: number
      required: [code, cartTotal]
    CouponValidation:
      type: object
      properties:
        valid:
          type: boolean
        discount:
          type: number
        type:
          type: ["string", "null"]
        message:
          type: string
      required: [valid, discount, message]
    CheckoutBody:
      type: object
      properties:
        couponCode:
          type: ["string", "null"]
      required: []
    CheckoutResult:
      type: object
      properties:
        success:
          type: boolean
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        totalCharged:
          type: number
        newBalance:
          type: number
      required: [success, orders, totalCharged, newBalance]
    Order:
      type: object
      properties:
        id:
          type: integer
        productName:
          type: string
        price:
          type: number
        status:
          type: string
          enum: [pending, delivered, cancelled]
        credentials:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, productName, price, status, createdAt]
    BuyProductBody:
      type: object
      properties:
        productId:
          type: integer
      required: [productId]
...
    ConvertPointsBody:
      type: object
      properties:
        points:
          type: integer
      required: [points]
    ConvertPointsResult:
      type: object
      properties:
        success:
          type: boolean
        pointsConverted:
          type: integer
        eurEarned:
          type: number
        newBalance:
          type: number
        newPoints:
          type: integer
      required: [success, pointsConverted, eurEarned, newBalance, newPoints]
...